+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * Added Flex support for custom media tasks
 1. [](#improved)
     * Added support for syslog and syslog facility logging (default: 'file')
+    * Improved usability of `System` configuration blueprint with side-tabs
 1. [](#bugfix)
     * Fixed asset manager to not add empty assets when they don't exist in the filesystem
     * Regression: Fixed asset manager methods with default legacy attributes