+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 1. [](#new)
     * Added `Utils::getMimeByFilename()`, `Utils::getMimeByLocalFile()` and `Utils::checkFilename()` methods
     * Added configurable dangerous upload extensions in `security.yaml`
-
+1. [](#improved)
+    * Updated vendor libraries to latest
+    
 # v1.5.2
 ## 10/01/2018